Make sure observermanager is initialized before RunObserver returns
The `ensureInitialized` function must be completed before `RunObserver` returns.  While removing some duplicate code, what appeared to be an extra `ensureInitialized` call was removed since it was already called inside of thread.  This caused a problem because `RunObserver` returned and allowed execution to continue before `ensureInitialize` was run inside thread.  Fix was to restore `ensureInitialize` call outside of thread